created js imports for custom.js
