+++ package.json
@@ -18,6 +18,7 @@
   "homepage": "https://github.com/NetoECommerce/Skeletal",
   "devDependencies": {
     "gulp": "3.9.0",
+    "gulp-js-import": "^1.0.6",
     "gulp-less": "3.0.3"
   },
   "dependencies": {
